% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{shape_badges}
\alias{shape_badges}
\title{Compute shape badges for a named list of dataframes}
\usage{
shape_badges(dfs)
}
\arguments{
\item{dfs}{Named list of dataframes.}
}
\value{
A named character vector of shape badges.
}
\description{
Compute shape badges for a named list of dataframes
}
\examples{
shape_badges(list(cars = cars, mtcars = mtcars))
}
