% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targets_available}
\alias{targets_available}
\title{Check whether a targets cache exists and targets is installed}
\usage{
targets_available(
  project_dir = ".",
  store = file.path(project_dir, "_targets")
)
}
\arguments{
\item{project_dir}{Path to the project directory.}

\item{store}{Path to the targets store (default \verb{_targets/} inside project).}
}
\value{
\code{TRUE} if targets is available and a cache is present.
}
\description{
Check whether a targets cache exists and targets is installed
}
